fix auth token http server used go func
